Finished merge script

+++ merge.sh
@@ -1,18 +1,34 @@
 #!/bin/bash
 
-mkdir -p _merged
-rm _merged/*
+program_exists () {
+	type "$1" &> /dev/null ;
+}
 
-for ((i = 1 ; i <= 99 ; i+=2 )); do
-    NUM1=$(printf '%03d' $i)
-    NUM2=$(printf '%03d' $((i+1)))
-    FILE1="${NUM1}.pdf"
-    FILE2="${NUM2}.pdf"
-    echo $FILE1 $FILE2
+main()
+{
+    if ! program_exists pdfjam; then
+        echo "Please install pdfjam"
+        exit 2
+    fi
 
-    pdfjam --angle 180 _out/$FILE2 --outfile _tmp/$FILE2
+    mkdir -p _merged
+    mkdir -p _tmp
+    rm _merged/*
+    rm _tmp/*
 
-    OUTFILE="${NUM1}-${NUM2}.pdf"
+    for ((i = 1 ; i <= 100 ; i+=2 )); do
+        NUM1=$(printf '%03d' $i)
+        NUM2=$(printf '%03d' $((i+1)))
+        FILE1="${NUM1}.pdf"
+        FILE2="${NUM2}.pdf"
+        echo $FILE1 $FILE2
 
-    pdfjam --nup 1x2 _out/$FILE1 _tmp/$FILE2 --outfile o.pdf
-done
+        pdfjam --angle 180 --fitpaper true --rotateoversize true _out/$FILE2 --outfile _tmp/$FILE2
+
+        OUTFILE="${NUM1}-${NUM2}.pdf"
+
+        pdfjam --nup 1x2 _out/$FILE1 _tmp/$FILE2 --outfile _merged/$OUTFILE
+    done
+}
+
+main "$@"
